Refactor: move getPythonApiCode to apiModal folder
